acc/backend/requirements.txt
flask
flask-sock
ultralytics
Pillow